move string formatters to their own folder
