+++ AttunedWeb/front-end/src/StringFormatters/getRatingString.ts
@@ -1,4 +1,4 @@
-import { Rating } from "./dtos/Rating.ts";
+import { Rating } from "../dtos/Rating.ts";
 
 export const getRatingString = (value: Rating) =>
   `${Array(value).fill("★").join("")}${Array(5 - value)
